refactor(milkscript): extract CLIcore_script_intercept loops into segmented component files
